Move UrlChecker to decoders package
